Refactor common parts out of ScheduledAsyncMethod
